+++ src/main/Main.hs
@@ -464,11 +464,16 @@ commandLineHandler currentDir progName isInterpreter = complicatedOptions
                           dockerCleanupOptsParser)
         addSubCommands'
             ConfigCmd.cfgCmdName
-            "Subcommands specific to modifying stack.yaml files"
-            (addCommand' ConfigCmd.cfgCmdSetName
-                        "Sets a field in the project's stack.yaml to value"
-                        (withConfig . cfgCmdSet)
-                        configCmdSetParser)
+            "Subcommands for accessing and modifying configuration values"
+            (do
+               addCommand' ConfigCmd.cfgCmdSetName
+                          "Sets a field in the project's stack.yaml to value"
+                          (withConfig . cfgCmdSet)
+                          configCmdSetParser
+               addCommand' ConfigCmd.cfgCmdEnvName
+                          "Print environment variables for use in a shell"
+                          (withConfig . withDefaultEnvConfig . cfgCmdEnv)
+                          configCmdEnvParser)
         addSubCommands'
           "hpc"
           "Subcommands specific to Haskell Program Coverage"