+++ test/integration/tests/620-env-command/files/Main.hs
@@ -1,0 +1,4 @@
+import Control.Concurrent.Async ()
+
+main :: IO ()
+main = pure ()